refactor: extract PendingTranscriptionFailureReason.swift from PendingTranscription.swift
Splits the `PendingTranscriptionFailureReason` enum (~113 lines, with
localized/tone/help-text accessors) out of PendingTranscription.swift
so that file holds only the small PendingTranscription value.

Byte-identical move of lines 3-115. Both files import Foundation.

PendingTranscription.swift: 146 → 32 lines. Tests: 667.
